requirements.txt
python-dotenv
prometheus_client
requests